chores: add input field component
