+++ package.json
@@ -14,9 +14,14 @@
     "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.11.16",
     "@mui/material": "^5.13.0",
+    "date-fns": "^2.30.0",
+    "dayjs": "^1.11.7",
+    "focus-trap-react": "^10.1.4",
     "react": "^18.2.0",
+    "react-day-picker": "^8.7.1",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.9",
+    "react-popper": "^2.3.0",
     "react-router": "^6.11.1",
     "react-router-dom": "^6.11.1"
   },
